Create not found, create constant Heroes
