+++ laura-sabater/lectures/redux/redux-lecture/src/components/HeroDetails/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function HeroDetails() {
+
+}
+
+export default HeroDetails;
