+++ laura-sabater/lectures/redux/redux-lecture/src/components/HeroesList/index.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+function HeroesList() {
+  const HEROES = [
+    { id: 11, name: 'Dr Nice' },
+    { id: 12, name: 'Narco' },
+    { id: 13, name: 'Bombasto' },
+    { id: 14, name: 'Celeritas' },
+    { id: 15, name: 'Magneta' },
+    { id: 16, name: 'RubberMan' },
+    { id: 17, name: 'Dynama' },
+    { id: 18, name: 'Dr IQ' },
+    { id: 19, name: 'Magma' },
+    { id: 20, name: 'Tornado' },
+  ];
+  return (
+    <div>
+      {HEROES.map((hero) => <li>{hero.name}</li>)}
+    </div>
+  );
+}
+
+export default HeroesList;
